Changed Widget constructor: added constructor to perform computation on the SFML object
This constructor accepts a callable object (function object, function pointer, lambda, ...) that receives a reference to graphicObject_ as paramater and performs computations on it.

Textures, parameters, and so on shall be members of the function object or captured by the lambda.

Note: if it is implemented as a function object, the user might save them object in a static member to internally have access to it.
This bypasses the wrapper functions (setPosition and so on) implemented to use the class.

Must find a way to prevent this behaviou.

Textures, parameters, and so on shall be members of the function object or captured by the lambda.

Note: if it is implemented as a function object, the user might save them object in a static member to internally have access to it.
This bypasses the wrapper functions (setPosition and so on) implemented to use the class.

Must find a way to prevent this behaviourr