Added wrapper class to safely manage the graphicObject_ inside function object called from the constructor to initialize the object
The Widget class allows the user to pass a callable object. This object can perform computation on the internal graphic Object, but must not access it directly. The function receives a reference to the wrapper object and must call the assign function on it in order to move-assign the graphic object to the widget.

Note: move assignment is mandatory, so the user won't have the private object available after a call to this function